[AFTF-23] refactor(NavBar): move component to ui
