+++ src/Components/common/App.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { AftRouter } from "../../routers/AftRouter";
+
+export const App = () => (
+  <>
+    <AftRouter />
+  </>
+);
