+++ src/Components/ui/navBar/NavBar.jsx
@@ -1,6 +1,6 @@
 import React, { useContext, useEffect } from "react";
 import { NavLink } from "react-router-dom";
-import { UserContext } from "../../contexts/UserContext";
+import { UserContext } from "../../../contexts/UserContext";
 import { checkToken } from "./checkToken";
 
 const NavBar = () => {
@@ -18,7 +18,7 @@ const NavBar = () => {
   return (
     <nav className="nav-bar">
       <h1>AFT</h1>
-      <div className="nav-bar_list">
+      <ul className="nav-bar_list">
         <NavLink to="/" className={activeLink}>
           🏠 Inicio
         </NavLink>
@@ -35,7 +35,7 @@ const NavBar = () => {
         <NavLink to="/login" className={activeLink}>
           🔐 Login
         </NavLink>
-      </div>
+      </ul>
     </nav>
   );
 };
