--- src/helpers/navBar/navBar.css
@@ -1,14 +1,0 @@
-.nav-bar {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 0.5rem 1rem;
-    background-color: var(--black-clean);
-    border-bottom: 1px solid var(--yellow);
-}
-
-.nav-bar_list {
-    display: flex;
-    align-items: center;
-    gap: 1rem;
-}
