+++ src/index.css
@@ -5,6 +5,7 @@
 
 :root {
   --red: rgb(150, 36, 36);
+  --pale-red: #c63906d2;
   --yellow: #b9a42e;
   --black-clean: #131313;
   --black-clean-opacity: #131313c6;
@@ -14,6 +15,12 @@
   --green: #1b891b;
   --purple: rebeccapurple;
   --blue: #0f6486;
+  --blue-gray: #5e7c9b;
+  --pale-blue: #c6d6f0;
+  --pale-green: #c6f0c6;
+  --pale-purple: #f0c6f0;
+  --pale-orange: #f0c6c6;
+  --pale-yellow: #f0c6c6;
 }
 
 .small {
@@ -35,6 +42,7 @@
 .blur {
   filter: blur(5px);
 }
+
 .action-button {
   width: 200px;
   height: 70px;
@@ -44,6 +52,12 @@
   padding: 0.5rem 1rem;
   transition: all 0.2s;
 }
+
+.action-button:hover {
+  background-color: var(--orange);
+  color: var(--white);
+}
+
 .bg-yellow {
   background-color: var(--yellow);
 }
@@ -55,6 +69,7 @@
 .bg-green {
   background-color: var(--green);
 }
+
 .bg-blue {
   background-color: var(--blue);
 }
@@ -76,6 +91,11 @@
   width: 100%;
   height: 100%;
   max-width: 300px;
+  object-fit: contain;
+  border-radius: 5px;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
+  transition: all 0.2s;
+  cursor: pointer;
 }
 
 .modal-content {
@@ -85,6 +105,7 @@
   width: 80%;
   max-width: 500px;
   border-radius: 5px;
+  transition: all 0.2s;
 }
 
 .modal-header {
@@ -109,19 +130,20 @@
   color: #aaaaaa;
   font-size: 28px;
   font-weight: bold;
+  transition: all 0.2s;
 }
 
 .close:hover {
-  box-shadow: inset 1px 2px 0 3px #c63906d2, 0px 0px 10px 0 rgba(210, 126, 23, 0.945);
+  box-shadow: inset 1px 2px 0 3px var(--pale-red), 0px 0px 10px 0 var(--orange);
 }
 
 p.success {
-  color: #1b891b;
+  color: var(--green);
 }
 
 p.error {
   margin-top: 0.5rem;
-  color: #cd4206;
+  color: var(--pale-red);
   font-size: small;
 }
 html,
@@ -130,6 +152,12 @@ body {
   color: var(--white);
   font-family: "Roboto", sans-serif;
   font-size: 1.2rem;
+  font-weight: bold;
+  line-height: 1.5;
+  margin: 0;
+  padding: 0;
+  text-align: center;
+  transition: all 0.2s;
 }
 
 a {
@@ -178,6 +206,9 @@ a {
 .card img {
   max-height: 170px;
   width: auto;
+  height: auto;
+  object-fit: contain;
+  border-radius: 5px;
 }
 
 .book-list {
@@ -210,6 +241,12 @@ button {
   box-shadow: inset 0 0 0 3px #675c123d, 0 6px 20px 0 rgba(0, 0, 0, 0.473);
 }
 
+button:hover {
+  background-color: var(--orange);
+  color: #fff;
+  box-shadow: inset 0 0 0 3px #675c123d, 0 6px 20px 0 var(--black-gray);
+}
+
 .portrait {
   width: 100px;
   height: 100px;
@@ -223,6 +260,12 @@ form {
   background-color: var(--black-clean-opacity);
   border-radius: 13px;
   box-shadow: 0 1px 20px rgba(177, 113, 16, 0.135), inset 0 0 0 3px #675c123d, 0 6px 20px 0 rgba(91, 83, 83, 0.111);
+  display: flex;
+  flex-direction: column;
+  justify-content: space-evenly;
+  align-items: center;
+  justify-items: stretch;
+  align-content: center;
 }
 
 input {
@@ -238,11 +281,15 @@ input {
   box-shadow: inset 0 0 0 3px #aa951f3d;
   width: 80%;
 }
+
 .content {
   padding: 2rem;
   display: grid;
   grid-template-rows: 3rem 1fr;
+  grid-template-columns: 1fr;
+  grid-template-areas: "header" "content";
 }
+
 .user-dashboard {
   display: flex;
   align-items: stretch;
@@ -250,11 +297,9 @@ input {
   color: aliceblue;
   font-size: 1.2rem;
   padding: 0.5rem 1rem;
-}
-
-.content_header {
-  padding: 0.5rem;
-  text-align: center;
+  background-color: var(--black-gray);
+  border-radius: 5px;
+  transition: all 0.2s;
 }
 
 .book {
@@ -291,18 +336,15 @@ input {
 .control-panel-mini button {
   cursor: pointer;
   max-width: 130px;
+  box-shadow: inset 0 0 0 3px #aa951f3d;
 }
 
 .panel {
   display: flex;
   flex-direction: column;
   align-items: center;
-  justify-content: space-between;
   width: 100%;
-  color: aliceblue;
-  font-size: 1.2rem;
-  padding: 0.5rem 1rem;
-  gap: 0.5rem;
+  background-color: var(--black-gray);
 }
 
 .flex-wrapper {
@@ -317,41 +359,30 @@ input {
 }
 
 .column {
+  display: flex;
   flex-direction: column;
-  align-items: center;
-  justify-content: space-between;
-  width: 100%;
   height: 100%;
-  color: aliceblue;
-  font-size: 1.2rem;
-  padding: 0.5rem 1rem;
-  gap: 0.5rem;
+}
+
+.row {
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-end;
 }
 
 .characters-list_item:hover {
-  background-color: var(--yellow);
+  background-color: var(--blue-gray);
   cursor: pointer;
 }
 
 .characters-list_item {
-  border: 1px solid red;
   display: flex;
   flex-direction: row;
   justify-content: space-evenly;
   padding: 0.5rem 1rem;
+  border-radius: 5px;
   background-color: var(--black-clean);
-}
-
-.d-row {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: flex-end;
-  width: 100%;
-  color: aliceblue;
-  font-size: 1.2rem;
-  padding: 0.5rem 1rem;
-  gap: 0.5rem;
+  border-radius: 5px;
 }
 
 .users-list {
@@ -360,9 +391,12 @@ input {
   align-items: stretch;
   list-style: none;
 }
+
 .users-list_item:hover {
   background-color: var(--yellow);
   cursor: pointer;
+  color: var(--black-gray);
+
 }
 
 .users-list_item {
