+++ src/index.js
@@ -1,14 +1,16 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
 import { UserProvider } from "./providers/UserProvider";
-import { AftRouter } from "./routers/AftRouter";
 import "./index.css";
+import { App } from "./Components/common/App";
 
 const divRoot = document.getElementById("root");
 const root = createRoot(divRoot);
 
 root.render(
-  <UserProvider>
-    <AftRouter />
-  </UserProvider>
+  <React.StrictMode>
+    <UserProvider>
+      <App />
+    </UserProvider>
+  </React.StrictMode>
 );
